📦 chore(Cargo.toml): add spinners crate 🔨 refactor(commit/mod.rs): remove debug println and add spinner while AI is generating commit message
The spinners crate has been added to the project to enable the display of spinner animations. In the commit/mod.rs file, a debug println statement has been removed. A spinner has been added to indicate that the AI is generating the commit message, and the spinner is stopped once the message has been generated.