🐛 fix(Cargo.toml): add lazy_static and async-recursion dependencies ✨ feat(commit command): add async_recursion to allow for recursive calls to run method ✨ feat(commit command): use Confirm dialoguer with a colorful theme 🐛 fix(commit command): use interact_opt instead of interact for Confirm dialoguer ✨ feat(utils): use lazy_static instead of structopt::lazy_static
The changes to Cargo.toml add two new dependencies, lazy_static and async-recursion. The lazy_static dependency is used in is_unicode_supported.rs to make the is_unicode_supported function a lazy static variable. The async_recursion dependency is used in commit/mod.rs to allow the run method to be called recursively.

In the commit command, a colorful theme is used with the Confirm dialoguer. The interact_opt method is used instead of interact to allow the user to exit the program if they don't want to stage and commit all files.

Finally, the structopt::lazy_static dependency in is_unicode_supported.rs has been replaced with the lazy_static dependency.