The changes to the `git.rs` file refactor the module to use a `GitRepository` struct that contains the different Git-related functions as methods. This makes it easier to manage the state of the Git repository, and also helps to ensure that the correct parameters are passed to each function.
The `assert_git_repo` function ensures that the current directory is a Git repository. The `get_changed_files` function gets a list of files that have been modified but not yet staged. The `get_staged_files` function gets a list of files that have been staged but not yet committed. The `get_staged_diff` function gets the diff for the files that have been staged.

The changes also include additional functionality to exclude lock files from `git diff`. The excluded files are printed out to `stderr` so that the user is aware of them.