The changes in this commit seem to be refactoring of the `run` function in the `commit` module.
The function now no longer uses the `async_recursion` macro and instead uses a loop to iterate through the code until a certain condition is met.

The condition in this case is whether there are any staged files or changed files. If there are no files, the program prompts the user to write some code and run the program again. If there are files, the program either stages all files or prompts the user to select the files to be staged.

The code also includes some minor formatting changes to the output.