The changes in the code include the addition of a new module `chat_context` in `src/commands/commit/mod.rs`. This module will contain the implementation for `ChatContext` struct, which will store the conversation context between the user and the autocommit feature.
The `get_chat_context` function is removed and replaced with the `ChatContext` struct. The `generate_autocommit_message` function now creates a new instance of `ChatContext` and adds the user message to it. The `generate_message` function is now a method of the `ChatContext` struct.

A new `i18n` module is created in `src/i18n/` directory, which contains the `Language` enum and the `I18n` and `I18nTranslation` structs. The `load_i18n` function loads translations for different languages into an instance of `I18n`. The `Language` enum currently contains only the `English` variant. The `commit messages` in the `I18nTranslation` struct have been hardcoded in English, but in the future, they can be translated based on the language selected by the user.