🎨 style(commit): format code using rustfmt 🚀 feat(commit): enable editing of commit message and ability to generate a new commit message
The rustfmt formatter was used to format the code. This commit also adds a feature to allow editing of the generated commit message and also generating a new commit message if the user chooses to. This is aimed at making the commit message more descriptive and precise. The user can now input any commit message they wish or opt to use a generated message.