It looks like a Git repository's `Cargo.toml` file and Rust source code for a `ChatContext` struct and `generate_message` function are being modified in this diff.
In the `Cargo.toml` file, a new dependency on `log` version 0.4.17 is added.

In the `ChatContext` struct, the `log::{debug, info}` module is imported. The `get_messages` method is changed to be public. The `get_initial_context` method is modified to use some variables from `config_data` instead of calling it each time, which may improve readability. The `generate_message` method is modified to log debug and info messages.

In the `mod.rs` file, there is an added blank line.