The changes to the Git autocommit tool include removing a file called `generate.rs` and moving its functions into the main `mod.rs` file. Additionally, a new `ChatContext` module was added to the imports. Finally, the `prompt.rs` file was also removed, and its functions were added to the `mod.rs` file.
The `generate.rs` file previously contained the `generate_autocommit_message` function which generated a commit message based on a provided configuration and the diff of the staged changes. This function was moved to the `CommitCommand` module in `mod.rs` and was renamed to `generate_autocommit_message`. This function now takes `&self` as the first argument since it's now a method of the `CommitCommand` struct.

The `ChatContext` module was added to the imports and is used in `mod.rs`. It was likely added to allow the `generate_autocommit_message` function to generate commit messages based on a conversation between a user and a chatbot.

The `prompt.rs` file was removed, and its two functions, `prompt_for_selected_files` and `prompt_to_commit_changes`, were moved to the `mod.rs` file. Both functions were updated to use `&self` as the first argument since they are now methods of the `CommitCommand` struct.

Two new functions were also added to the `CommitCommand` struct. The `prompt_for_remote` function prompts the user for a remote Git repository, while the `prompt_for_push` function prompts the user to confirm if they want to push changes to the specified remote.