🎉 feat(git): Use `git2` library to get remotes and checkout new branch
This commit replaces the usage of the `Command` module with the `git2` library to get the list of remotes in the repository. It also uses the `target()` method of `git2::Reference` to get the target of HEAD instead of unwrapping the result of `head()` method. Additionally, the method `git_checkout_new_branch()` now uses the `git2` library to checkout a new branch.

Finally, minor code formatting improvements were made in the `GitRepository::get_statuses()` method.