🐛 fix(git/mod.rs): Fix output formatting for directory skipping ✨ feat(git/mod.rs): Add check for uncommitted changes before committing In the `GitRepository` implementation, the output formatting was updated for when a directory is skipped during the `add_all_files_to_index` method. Previously, the "Skipping directory" message was preceded by a right arrow emoji, which was confusing. The arrow emoji was removed and the message now starts with "Skipping directory" without any additional symbol.
Also, a new feature was added to check if there are any uncommitted changes in the repository before committing the changes made by the script. The `statuses` method from the `git2::Repository` struct is used to get the repository's status, and if there are any changes, an error message will be returned. The error message will include all uncommitted changes with their status and path.

Changes made:
- The output message for skipped directories was updated to remove the right arrow emoji and make the message clear.
- A new check was added to the `get_commit_summary_table` method to see if there are any uncommitted changes before committing. If there are any uncommitted changes, an error message will be returned with a list of all the uncommitted changes and their status.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>